Use original tag string where appropriate
When we parse a version string like `1.0-foo` which lacks the semver patch
field, everything works fine in terms of sorting and filtering versions
because the semver parser normalizes it to `1.0.0-foo`.

But when we use the normalized version string to resolve the image digest,
and propagate it as the `tag` in the resource, we end up with "Not Found"
errors.

This change ensures the original tag string is used for everything except
sorting and comparing versions.